+++ addons/dialogic/Editor/ThemeEditor/ThemeEditor.gd
@@ -88,10 +88,12 @@ onready var n : Dictionary = {
 	'name_shadow_offset_x': $"VBoxContainer/TabContainer/Name Label/Column/GridContainer/HBoxContainer/ShadowOffsetX",
 	'name_shadow_offset_y': $"VBoxContainer/TabContainer/Name Label/Column/GridContainer/HBoxContainer/ShadowOffsetY",
 	'name_bottom_gap': $"VBoxContainer/TabContainer/Name Label/Column3/GridContainer/HBoxContainer5/BottomGap",
+	'name_horizontal_offset': $"VBoxContainer/TabContainer/Name Label/Column3/GridContainer/HBoxContainer5/HorizontalOffset",
 	'name_background_modulation': $"VBoxContainer/TabContainer/Name Label/Column2/GridContainer/HBoxContainer6/CheckBox",
 	'name_background_modulation_color': $"VBoxContainer/TabContainer/Name Label/Column2/GridContainer/HBoxContainer6/ColorPickerButton",
 	'name_padding_x': $"VBoxContainer/TabContainer/Name Label/Column2/GridContainer/HBoxContainer/NamePaddingX",
 	'name_padding_y': $"VBoxContainer/TabContainer/Name Label/Column2/GridContainer/HBoxContainer/NamePaddingY",
+	'name_position': $"VBoxContainer/TabContainer/Name Label/Column3/GridContainer/HBoxContainer/Positions",
 	
 	
 	# Choice Buttons
@@ -162,6 +164,15 @@ func _ready() -> void:
 	n['button_hover'].connect('style_modified', self, '_on_choice_style_modified')
 	n['button_pressed'].connect('style_modified', self, '_on_choice_style_modified')
 	n['button_disabled'].connect('style_modified', self, '_on_choice_style_modified')
+	
+	n['name_position'].text = 'Left'
+	n['name_position'].connect('item_selected', self, '_on_name_position_selected')
+	var name_positions_popup = n['name_position'].get_popup()
+	name_positions_popup.clear()
+	name_positions_popup.add_radio_check_item('Left')
+	name_positions_popup.add_radio_check_item('Center')
+	name_positions_popup.add_radio_check_item('Right')
+	n['name_position'].select(0)
 	
 	
 	# Character Picker
@@ -302,6 +313,9 @@ func load_theme(filename):
 	n['name_shadow_offset_x'].value = theme.get_value('name', 'shadow_offset', Vector2(2,2)).x
 	n['name_shadow_offset_y'].value = theme.get_value('name', 'shadow_offset', Vector2(2,2)).y
 	n['name_bottom_gap'].value = theme.get_value('name', 'bottom_gap', 48)
+	n['name_horizontal_offset'].value = theme.get_value('name', 'horizontal_offset', 0)
+	
+	n['name_position'].select(theme.get_value('name', 'position', 0))
 	
 	
 	# Next indicator animations
@@ -369,11 +383,12 @@ func _on_PreviewButton_pressed() -> void:
 	var preview_dialog = Dialogic.start('', true, "res://addons/dialogic/Dialog.tscn", false, false)
 	preview_dialog.preview = true
 	
-	if n['character_picker'].text == 'Random Character':
-		var characters : Array = DialogicUtil.get_character_list()
-		if characters.size():
-			characters.shuffle()
-			preview_character_selected = characters[0]['file']
+	if n['character_picker']: # Sometimes it can't find the node
+		if n['character_picker'].text == 'Random Character':
+			var characters : Array = DialogicUtil.get_character_list()
+			if characters.size():
+				characters.shuffle()
+				preview_character_selected = characters[0]['file']
 
 	preview_dialog.dialog_script = {
 			"events":[
@@ -750,9 +765,17 @@ func _on_name_padding_value_changed(_value) -> void:
 func _on_name_BottomGap_value_changed(value) -> void:
 	if loading:
 		return
-	DialogicResources.set_theme_value(current_theme, 'name', 'bottom_gap', value)
+	# Todo in 2.0: Replace for a single Vector2 instead of two variables
+	DialogicResources.set_theme_value(current_theme, 'name', 'bottom_gap', n['name_bottom_gap'].value)
+	DialogicResources.set_theme_value(current_theme, 'name', 'horizontal_offset', n['name_horizontal_offset'].value)
 	_on_PreviewButton_pressed() # Refreshing the preview
 
+
+func _on_name_position_selected(index):
+	if loading:
+		return
+	DialogicResources.set_theme_value(current_theme, 'name', 'position', index)
+	_on_PreviewButton_pressed() # Refreshing the preview
 
 ## ------------ 		CHOICE BUTTON TAB	 	--------------------------------